Update location of svg-icon (move to own util file)
