+++ js/util/svg-icon.ts
@@ -1,0 +1,1 @@
+export type SvgIcon = React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
